fix: prevent terminal flashing during agent streaming
Colocate stats subscriptions with tab label components to avoid
re-rendering the entire RightSidebarTabsetNode tree when stats
update during agent activity.
